+++ CHANGELOG.md
@@ -2,12 +2,16 @@
 
 # 1.4.0 (edge, unreleased)
 
-* Expermental new features
+* Experimental new features
   * Computation setter functions now receive a context argument, and both getter and setter functions now receive a keypath argument.
   * Component styles can be split into multiple managed style tags in the document head by setting `Ractive.perComponentStyleElements` to `true`.
   * Methods that set a single value will now return the new value as the resolution of the returned promise e.g. `r.toggle('foo')` if `foo` is falsey will return a promise that resolves to `true`. `add`/`subtract`/`set` behave similarly.
   * Decorators and custom events can now access their local context, in the case of a yielded attribute partial, with `this.getLocalContext()`.
 
+* Typings (probably should be added to 1.3.x and 1.2.x)
+  * Add `Ractive.tick` definition
+  * Add `ListenerContextHelper` as new type which is the Context passed to ListenerCallback (includes additional `name` property)
+  * Revert change to `.extend()` done with [#3349](https://github.com/ractivejs/ractive/pull/3349)
 
 # 1.3.14